+++ assets/plugin/index.js
@@ -1559,7 +1559,7 @@ function createMoonContextEngine(api) {
     info: {
       id: "moon",
       name: "Moon Context Engine",
-      version: "1.2.5",
+      version: "1.2.6",
       ownsCompaction: true,
     },
     bootstrap(params) {